refactor(workflow): fold control-plane helpers into service
Summary:
- move control-plane storage helpers into workflow service.go
- delete the separate control_plane_store.go helper file

Rationale:
- satisfy the architecture guard by keeping Ent-backed helpers on the
  existing allowed service boundary

Tests:
- make lint-architecture
- PATH=/home/yuzhong/.local/go1.26.1/bin:$PATH go test ./internal/workflow/... ./internal/httpapi/... ./internal/orchestrator/...
